POP3 deletes view update should be immediate
Currently POP3 view can "lag" behind due to delays of the mailbox event processing
(as a reminder asynchronous), which thus causes QUIT to leave (temporarily)
metadata without associated messages to subsequent sessions.

Doing an aggressive delete to the view in the POP3 transaction allows to stay consistent.
We want to keep event bus retries to leverage its reliability in case of partial execution.

This could be reproduced by adding a delay in PopulateMetadataStoreListener::handleExpunged
(100 ms).